requirements.txt
Django==4.1.4
djangorestframework==3.14.0
plaid==0.1.7
plaid_python==11.3.0
python-dotenv==1.0.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-dotenv==22.3.0
